first check-in
This is the first check-in for this repository.
Contains yocto project updated for gcc 8.2,lttng 2.11 that fixed an issue for .net core 3.1.1, .net core 3.1.1 meta layer, mono meta layer needed for .net core, a fixt for xmlsec1 for download path , compulab meta leyers, qt5.14.1 meta layer. Also the conf folder that contains an local.conf and bblayers.conf updated to include all packages that we need.